requirements.txt
streamlit>=1.28.0
firecrawl-py>=0.0.8
openai>=1.3.0,<2.0.0
googlesearch-python>=1.2.3
pandas>=2.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tenacity>=8.2.0
aiohttp>=3.8.0
asyncio-throttle>=1.0.0
python-dateutil>=2.8.0
pytz>=2023.3
lxml>=4.9.0
html5lib>=1.1
httpx>=0.24.0